Moved feature.hpp to depend on geometry.hpp and not vise versa
