+++ include/mapbox/geometry.hpp
@@ -7,7 +7,6 @@
 #include <mapbox/geometry/multi_line_string.hpp>
 #include <mapbox/geometry/multi_polygon.hpp>
 #include <mapbox/geometry/geometry.hpp>
-#include <mapbox/geometry/feature.hpp>
 #include <mapbox/geometry/point_arithmetic.hpp>
 #include <mapbox/geometry/for_each_point.hpp>
 #include <mapbox/geometry/envelope.hpp>
