+++ tests/test.cpp
@@ -1,8 +1,10 @@
 #include <mapbox/geometry.hpp>
+#include <mapbox/feature.hpp>
 
 #include <cassert>
 
 using namespace mapbox::geometry;
+using namespace mapbox::feature;
 
 static void testPoint() {
     point<double> p1;
